backend/brainwave_backend/requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.1.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.6
django-cors-headers==4.4.0
django-positions==0.6.0
django-sequences==3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docstring_parser==0.16
frozenlist==1.4.1
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.62.2
gunicorn==23.0.0
idna==3.7
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.12
langchain-core==0.2.28
langchain-google-vertexai==1.0.8
langchain-text-splitters==0.2.2
langsmith==0.1.96
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.6
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.4
psycopg2==2.9.10
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==72.2.0
shapely==2.0.5
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.0
tenacity==8.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
